ai-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
thefuzz
httpx
